Extract out default story
